// tradesmath-frontend/src/app/models/questions.ts
export class Questions {
    key: string | undefined;
    questionText: string | undefined;
    equation: string | undefined;
    min1: string | undefined;
    max1: string | undefined;
    step1: string | undefined;
    min2: string | undefined;
    max2: string | undefined;
    step2: string | undefined;
    topicId: string | undefined;
    imageLink: string | undefined;
    videoLink: string | undefined;
}